requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.2.0
discord.py==2.3.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
yarl==1.9.2
